+++ ContextCache/changelog.md
@@ -1,5 +1,8 @@
 ﻿# Changelog
+## 1.1.0 (2025-06-06)
+ - Added `Get-ContextCache`
+ - Added `Tab-Expansion`
+ - BREAKING `CacheKey` Param renamed to `Name`
+ - BREAKING `Module` Param removed
 ## 1.0.0 (2025-06-05)
- - New: Some Stuff
- - Upd: Moar Stuff
- - Fix: Much Stuff
+ - First version